Dont Delete Entities which are nott in data.xml file
Hi Wael,
We are using your extract config migration data task in Azure Devops pipeline. I observe that you are deleting the entities data which are not there in data.xml file. Can you comment these particular lines as we need all the entities data xml files which are created prior to the execution of the current even though the entity is not present in the data.xml file which is be exported using CMT tool. We are using this task and pushing the changes to our github for tracking of the files or entities xml's. So , by commenting these lines we will be able to retain the files and can use for source tracking inside our GitHub.
Need your help in this scenario.
If you have any other option to avoid deleting the previous files using your tasks, that's also fine for us to proceed.
Thank you.